Fix: Code climate improovement. Function ModalScreen has 36 lines of code (exceeds 25 allowed). Consider refactoring.
