+++ src/container/login/modal.container.js
@@ -1,50 +1,35 @@
-import React, {useEffect} from 'react';
-import auth from '@react-native-firebase/auth';
+import React from 'react';
 import {
   faCheckCircle,
   faSpinner,
   faTimesCircle,
 } from '@fortawesome/free-solid-svg-icons';
-import {StackActions} from '@react-navigation/routers';
 import {ModalLayout} from '../../components/login';
+import {useRedirect} from '../../library/hooks';
 const ModalScreen = props => {
-  useEffect(() => {
-    const timeout = setTimeout(() => {
-      if (auth()?.currentUser?.uid) {
-        props.navigation.dispatch(StackActions.replace('BookingList'));
-      }
-    }, 3000);
-    return () => {
-      clearTimeout(timeout);
-    };
-  }, [auth().currentUser]);
-  const {data, error, loading} = props;
-  if (data) {
-    return (
-      <ModalLayout
-        icon={faCheckCircle}
-        text={props.login ? 'Logged In' : 'Signed Up'}
-      />
-    );
-  }
+  useRedirect(props);
+  const {data, error, loading, login} = props;
 
-  if (error) {
-    return (
-      <ModalLayout
-        icon={faTimesCircle}
-        text={props.login ? "Can't Log in" : "Can't sign up"}
-      />
-    );
-  }
-  if (loading) {
-    return (
-      <ModalLayout
-        icon={faSpinner}
-        text={props.login ? 'Logging In...' : 'Signing up...'}
-      />
-    );
-  }
-  return <></>;
+  const newProps = () => {
+    if (data) {
+      return {icon: faCheckCircle, text: login ? 'Logged In' : 'Signed Up'};
+    } else if (error) {
+      return {
+        icon: faTimesCircle,
+        text: login ? "Can't Log in" : "Can't sign up",
+      };
+    } else if (loading) {
+      return {
+        icon: faSpinner,
+        text: login ? 'Logging In...' : 'Signing up...',
+      };
+    }
+    return '';
+  };
+
+  if (!newProps) return <></>;
+
+  return <ModalLayout {...newProps} />;
 };
 
 export default ModalScreen;
